trying to get the IAs to run
