+++ saviSimulation/savi/asl/working.asl
@@ -83,7 +83,7 @@ targetFar :-
 	:	target(threat,TYPE,AZ,EL,RANGE) & 
 		position(X_REF,Y_REF,Z_REF,_) &
 		velocity(BEARING,_,_,_)
-	<-	.GetAbsolutePosition(X_TARGET,Y_TARGET,Z_TARGET,X_REF,Y_REF,Z_REF,BEARING,AZ,EL,RANGE)
+	<-	UxVInternalActions.GetAbsolutePosition(X_TARGET,Y_TARGET,Z_TARGET,X_REF,Y_REF,Z_REF,BEARING,AZ,EL,RANGE)
 		.broadcast(tell,threatSeen(X_TARGET,Y_TARGET,Z_TARGET)).
 
 // Default plan for observing target - force recursion.
